chore: Refactor Memory Pools
