+++ native/core/src/execution/mod.rs
@@ -29,9 +29,7 @@ pub(crate) mod util;
 pub use datafusion_comet_spark_expr::timezone;
 pub(crate) mod utils;
 
-mod fair_memory_pool;
-mod memory_pool;
-pub use memory_pool::*;
+mod memory_pools;
 
 #[cfg(test)]
 mod tests {
